# Conflicts:
#	app/src/main/res/values-phi/strings.xml
#	app/src/main/res/values-vi/strings.xml
#	app/src/main/res/values-zh/strings.xml
#	app/src/main/res/values/strings.xml

+++ app/src/main/java/org/cxct/sportlottery/network/bet/settledList/MatchOdd.kt
@@ -25,7 +25,7 @@ data class MatchOdd(
     @Json(name = "matchId")
     val matchId: String?,
     @Json(name = "odds")
-    val odds: Int?,
+    val odds: Float?,
     @Json(name = "oddsId")
     val oddsId: String?,
     @Json(name = "playCateCode")
